+++ lib/rails-bootstrap-helpers/rails/engine.rb
@@ -2,6 +2,7 @@ module RailsBootstrapHelpers
   module Rails
     class Engine < ::Rails::Engine
       initializer "rails-bootstrap-helpers.helpers" do
+        ActionView::Base.send :include, RailsBootstrapHelpers::Helpers::AccordionHelper
         ActionView::Base.send :include, RailsBootstrapHelpers::Helpers::AlertHelper
         ActionView::Base.send :include, RailsBootstrapHelpers::Helpers::BaseHelper
         ActionView::Base.send :include, RailsBootstrapHelpers::Helpers::ButtonHelper